ti-j72xx: viddec-test-app: initialize string array to zero
Issue: LIN1021-2482

In order to fix the failure "Failed to open drm device", the reason is
that there is some mess values in the array of the code in some situations,
cause the case judgment to fail and enter an illegal branch to run.

And add the macro definition "DRMMODE & DRM" to build the code.

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>